reorganize files in project
