+++ index.js
@@ -12,11 +12,11 @@ const accountRouter = require('./router/accounts/accountRouter');
 const investmentRouter = require('./router/investments/investmentRouter');
 const assetTypeRouter = require('./router/accounts/assetTypeRouter');
 const otherQuotesRouter = require('./router/investments/otherQuotesRouter');
-const creditCardRouter = require('./router/creditCardRouter');
-const expenseCreditCardRouter = require('./router/expensesCreditCardRouter');
-const mercadoPagoRouter = require('./router/mercadoPagoRouter')
-const periodRouter = require('./router/periodRouter')
-const refundRouter = require('./router/refundRouter')
+const creditCardRouter = require('./router/creditCard/creditCardRouter');
+const expenseCreditCardRouter = require('./router/creditCard/expensesCreditCardRouter');
+const mercadoPagoRouter = require('./router/periods/mercadoPagoRouter')
+const periodRouter = require('./router/periods/periodRouter')
+const refundRouter = require('./router/registers/refundRouter')
 const userRouter = require('./router/user/userRouter')
 const monthRegisterRouter = require('./router/registers/monthRegisterRouter')
 const auth = require('./middlewares/auth');
